Fix harmless signed/unsigned comparison warning in a test
Add a cast to avoid a warning recently added in 1f40a7e4e3 (Document
that wxDataViewCtrl::SetSelections() skips invalid items, 2020-10-12).

+++ tests/controls/dataviewctrltest.cpp
@@ -154,7 +154,7 @@ TEST_CASE_METHOD(MultiSelectDataViewCtrlTestCase,
     REQUIRE_NOTHROW( m_dvc->SetSelections(sel) );
 
     wxDataViewItemArray sel2;
-    CHECK( m_dvc->GetSelections(sel2) == sel.size() );
+    CHECK( m_dvc->GetSelections(sel2) == static_cast<int>(sel.size()) );
 
     CHECK( sel2 == sel );
